+++ src/main/java/chess/web/dto/GameStateDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.web.dto;
 
 import chess.domain.Color;
 import chess.domain.board.Board;
